WIP - evolution
Refactor the whole application for the new concepts:
- Backends
- Notifiables

So the idea is that it can support different backends (PD, VectorOps,
..etc)

While at the same time it can notify any client that is notifiable.
Currently it notifies a client called Cui that used to be the main
application before, which was the command line interface.